+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # FastReseg 1.1.2
 
-* Add defensive code for dimension change during data subset in `get_neighborhood_content()`.
+* Add defensive code to convert input `refProfiles` as base matrix in `runPreprocess()`.
 * Add check for xy colinearity edge case in `createSpatialDelaunayNW_from_spatLocs()`.
 * Add fallback transcript grouping for `groupTranscripts_Delaunay()` in case of failed Delaunay.